fix:Entity PkType 수정

+++ backend/src/main/java/com/fxflow/domain/mockbankaccount/repository/MockBankAccountRepository.java
@@ -3,5 +3,5 @@ package com.fxflow.domain.mockbankaccount.repository;
 import com.fxflow.domain.mockbankaccount.entity.MockBankAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MockBankAccountRepository extends JpaRepository<MockBankAccount,Integer> {
+public interface MockBankAccountRepository extends JpaRepository<MockBankAccount,Long> {
 }
